tests/wpt/web-platform-tests/navigation-api/scroll-restoration/manual-restoreScroll-repeated.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<div style="height: 1000px; width: 1000px;"></div>
<div id="frag"></div>
<script>
promise_test(async t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));
  assert_equals(window.scrollY, 0);
  await navigation.navigate("#frag").finished;
  assert_not_equals(window.scrollY, 0);
  navigation.onnavigate = t.step_func(e => {
    e.transitionWhile(Promise.resolve(), { scrollRestoration: "manual" });
    e.restoreScroll();
    assert_equals(window.scrollY, 0);
    assert_throws_dom("InvalidStateError", () => e.restoreScroll());
  });
  await navigation.back().finished;
  assert_equals(window.scrollY, 0);
}, "scrollRestoration: restoreScroll() should throw if called a second time");
</script>
</body>
